removed experiment script

+++ cleartext/pipeline/pipeline.py
@@ -210,7 +210,7 @@ class Pipeline(object):
         # run model with dummy target
         sources, targets = zip(*((example.src, example.trg) for example in self.test_data[:num_examples]))
         source_tensor = self.src.process(sources).to(self.device)
-        _dummy = torch.zeros(source_tensor.shape, dtype=int, device=self.device)
+        _dummy = torch.zeros(source_tensor.shape, dtype=torch.int, device=self.device)
         _dummy.fill_(self.trg.vocab[self.SOS_TOKEN])
 
         # select most likely tokens (ignoring non-word tokens)
